% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_from_izmir.R
\name{get_data_from_izmir}
\alias{get_data_from_izmir}
\title{get_data_from_izmir}
\usage{
get_data_from_izmir(data_name)
}
\arguments{
\item{data_name}{A data set name obtained from list_data_names()}
}
\value{
A data frame that shows the observations for given data name
}
\description{
Given a data name, this function returns the data frame
}
\examples{

get_data_from_izmir("asik-veysel-rekreasyon-alani-buz-pisti-kullanici-verileri")


}
